refactor: 폴더 구조 변경 및 appstae 위치 변경
